chore(tooling): relocate source scripts outside build roots and disambiguate ADR numbering
