+++ backend/api/nodemon.json
@@ -5,8 +5,7 @@
     "ext": "ts",
     "ignore": [
         "node_modules",
-        "dist",
-        "src/scripts"
+        "dist"
     ],
     "exec": "ts-node --files -r tsconfig-paths/register src/index.ts",
     "signal": "SIGTERM"
