+++ backend/api/package.json
@@ -15,17 +15,15 @@
     "lint": "eslint .",
     "lint:fix": "eslint src --ext .ts --fix",
     "db:seed": "node --max-old-space-size=2048 -r ts-node/register/transpile-only -r tsconfig-paths/register src/seeds/runSeeds.ts",
-    "seed:admin": "ts-node -r tsconfig-paths/register src/scripts/seed-admin.ts",
-    "seed:lifecycle": "ts-node -r tsconfig-paths/register src/scripts/verify-lifecycle-seed.ts",
-    "verify:job": "ts-node -r tsconfig-paths/register src/scripts/verify-job-trigger.ts",
-    "verify:logs": "ts-node -r tsconfig-paths/register src/scripts/verify-audit-logs.ts",
-    "verify:notifs": "ts-node -r tsconfig-paths/register src/scripts/verify-notifications.ts",
-    "verify:macherla": "ts-node -r tsconfig-paths/register src/scripts/verify-macherla-live.ts",
+    "seed:lifecycle": "ts-node -r tsconfig-paths/register scripts/verify-lifecycle-seed.ts",
+    "verify:job": "ts-node -r tsconfig-paths/register scripts/verify-job-trigger.ts",
+    "verify:logs": "ts-node -r tsconfig-paths/register scripts/verify-audit-logs.ts",
+    "verify:notifs": "ts-node -r tsconfig-paths/register scripts/verify-notifications.ts",
+    "verify:macherla": "ts-node -r tsconfig-paths/register scripts/verify-macherla-live.ts",
     "smoke": "node dist/smoke.js",
-    "smoke:fixtures": "ts-node -r tsconfig-paths/register src/scripts/ensure-listing-smoke-fixtures.ts",
-    "backup": "ts-node src/scripts/backup-database.ts",
-    "restore": "ts-node src/scripts/restore-database.ts",
-    "verify-backup": "ts-node src/scripts/verify-backup.ts",
+    "smoke:fixtures": "ts-node -r tsconfig-paths/register scripts/ensure-listing-smoke-fixtures.ts",
+    "backup": "npm run backup -w @esparex/core",
+    "verify-backup": "npm run verify-backup -w @esparex/core",
     "test": "cross-env NODE_NO_WARNINGS=1 jest --runInBand",
     "test:unit": "cross-env NODE_NO_WARNINGS=1 jest --testPathIgnorePatterns=src/__tests__/integration",
     "test:integration": "cross-env NODE_NO_WARNINGS=1 jest --testPathPatterns=integration",
@@ -34,16 +32,6 @@
     "test:watch": "cross-env NODE_NO_WARNINGS=1 jest --watch",
     "test:coverage": "cross-env NODE_NO_WARNINGS=1 jest --coverage",
     "test:ci": "cross-env NODE_NO_WARNINGS=1 jest --coverage --passWithNoTests --ci",
-    "ops": "ts-node -r tsconfig-paths/register src/scripts/ops/index.ts",
-    "ops:geo-repair": "npm run ops -- geo-repair",
-    "ops:location-audit": "npm run ops -- location-coverage-audit",
-    "ops:location-status-backfill": "npm run ops -- location-status-backfill",
-    "ops:report-unify": "npm run ops -- report-unify-backfill",
-    "ops:report-orphans": "npm run ops -- report-orphan-remediate",
-    "migrate:ad-listingtype-backfill": "ts-node -r tsconfig-paths/register src/scripts/migrate_ad_listingtype_backfill.ts",
-    "migrate:spare-part-listingtype-backfill": "ts-node -r tsconfig-paths/register src/scripts/migrations/migrate_spare_part_listing_types.ts",
-    "migrate:spare-part-needs-review-cleanup": "ts-node -r tsconfig-paths/register src/scripts/migrations/remove_spare_part_needs_review.ts",
-    "migrate:lifecycle-integrity-sweep": "ts-node -r tsconfig-paths/register src/scripts/migrations/migrate_listing_lifecycle_integrity.ts",
     "scan:master-data-integrity": "node scripts/master-data-integrity-scan.js",
     "validate:master-catalog-health": "node scripts/validate-master-catalog-health.js",
     "ops:indexes:apply": "ts-node scripts/syncUserIndexes.ts && ts-node scripts/sync_ad_indexes.ts",
