+++ backend/api/scripts/verify-job-trigger.ts
