+++ backend/api/scripts/verify-lifecycle-seed.ts
