+++ backend/api/scripts/verify-notifications.ts
