+++ core/package.json
@@ -50,7 +50,9 @@
     "clean": "node ../scripts/clean.js --core",
     "test": "jest",
     "type-check": "npm run build -w @esparex/shared && tsc --noEmit",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "backup": "ts-node scripts/backup-database.ts",
+    "verify-backup": "ts-node scripts/verify-backup.ts"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.954.0",
